+++ tests/test_visitors/test_decorators/test_alias_decorator.py
@@ -5,10 +5,10 @@ from wemake_python_styleguide.visitors.decorators import alias
 
 class _HasAliasedProp(object):
     def existing(self):
-        return None
+        """Existing."""
 
     def first(self):
-        return None
+        """First."""
 
 
 def test_raises_for_duplicates():
@@ -18,7 +18,7 @@ def test_raises_for_duplicates():
 
 
 def test_useless_alias():
-    """Ensures that decorator raises an exception for duplicates."""
+    """Ensures that decorator raises an exception for useless alias."""
     with pytest.raises(ValueError, match='duplicate'):
         alias('name', ('name',))