+++ wemake_python_styleguide/compat/nodes.py
@@ -10,8 +10,8 @@ import sys
 from typing import Any, Optional
 
 if sys.version_info >= (3, 8):  # pragma: py-lt-38
-    from ast import NamedExpr as NamedExpr  # noqa: WPS113, WPS433
     from ast import Constant as Constant  # noqa: WPS433, WPS113
+    from ast import NamedExpr as NamedExpr  # noqa: WPS113, WPS433
 else:  # pragma: py-gte-38
     class NamedExpr(ast.expr):  # noqa: WPS440
         """